Address review: rename _flyer.py to _flyable.py, inline watcher helper (#1241)
- Rename core/_flyer.py to core/_flyable.py (it holds StandardFlyable/
  FlyableLogic), updating the __init__ import and conf.py obj_ignore.
- Inline the observe_watcher_updates helper back into StandardMovable.set
  and StandardFlyable.complete; the two-line observe/yield loop didn't
  earn a shared function.
- Drop the plan-usage snippet from the write-a-flyer how-to; a full plan
  example will come in a later PR.
