+++ docs/conf.py
@@ -154,8 +154,8 @@ obj_ignore = [
     "ophyd_async.core._soft_signal_backend.Setter",
     "ophyd_async.core._status.AsyncStatusBase",
     "ophyd_async.core._standard_base._StandardBase",
-    "ophyd_async.core._flyer.PrepareT",
-    "ophyd_async.core._flyer.CtxT",
+    "ophyd_async.core._flyable.PrepareT",
+    "ophyd_async.core._flyable.CtxT",
     "ophyd_async.core._utils.P",
     "ophyd_async.core._utils.P.args",
     "ophyd_async.core._utils.P.kwargs",
